Removed all references to deprecated u_ types, which are apparently 	just left over from BSD and nobody wants to support them. 	This was mainly triggered by trying to compile with dietlibc. As a result of this, all integral-related functions, such as operator<<>> 	will take sized types only. It really looks more logical this way. s[io]stream will now read/write wchar_t as utf8, not as an int.

+++ bvt/bvt05.cc
@@ -59,7 +59,7 @@ void TestBigFill (const size_t size, const T magic)
 	cout << "works";
     else {
 	cout << "does not work: mismatch at " << distance(vbig.begin(), iMismatch);
-	cout << ", value = 0x" << ios::hex << u_long(*iMismatch) << ios::dec;
+	cout << ", value = 0x" << ios::hex << uintptr_t(*iMismatch) << ios::dec;
     }
     cout << endl;
 }
@@ -77,9 +77,9 @@ void TestBigCopy (const size_t size, const T magic)
 	cout << "works";
     else {
 	cout << "does not work: mismatch at " << distance(vbig1.begin(), iMismatch.first);
-	cout << ios::hex << ", 0x" << u_long(*iMismatch.first);
+	cout << ios::hex << ", 0x" << uintptr_t(*iMismatch.first);
 	assert (iMismatch.second < vbig2.end());
-	cout << " != 0x" << u_long(*iMismatch.second) << ios::dec;
+	cout << " != 0x" << uintptr_t(*iMismatch.second) << ios::dec;
     }
     cout << endl;
 }
